Refactor component exports for consistency
Standardized component export patterns across the project to named exports, improving consistency and enabling better tree-shaking optimization. Adopted the 'index.ts' barrel for cleaner imports. Additionally, renamed 'pass-words' to 'PassWords' for component naming consistency and capitalized file names to adhere to naming conventions. Commented out 'jsxImportSource' in tsconfig as it may no longer be necessary with the updated export strategy, but requires further testing.
