+++ src/components/CodePen.tsx
@@ -26,7 +26,7 @@ const Pen: React.FC<PenProps> = ({
   />
 );
 
-export default Pen;
+export { Pen };
 
 export const query = graphql`
   fragment Pen on ContentfulCodePen {
