+++ src/components/Tags.tsx
@@ -104,4 +104,4 @@ const Tags: React.FC = () => {
   );
 };
 
-export default Tags;
+export { Tags };
