+++ src/components/imageHeader.tsx
@@ -95,7 +95,7 @@ const HeaderImage = styled(GatsbyImage)`
   z-index: 1;
 `;
 
-const Header: React.FC = () => {
+const ImageHeader: React.FC = () => {
   const [selectedImage, setSelectedImage] = useState<HeaderImageNode | null>(
     null,
   );
@@ -172,4 +172,4 @@ const Header: React.FC = () => {
   );
 };
 
-export default Header;
+export { ImageHeader };
