+++ src/components/index.ts
@@ -1,12 +1,14 @@
-import ExternalLink from './ExternalLink';
-import Header from './header';
-import Layout from './layout';
-import ListTags from './ListTags';
-import Pen from './CodePen';
-import Posts from './posts';
-import Seo from './seo';
-import SiteMap from './sitemap';
-import Tags from './Tags';
+import { ExternalLink } from './ExternalLink';
+import { Header } from './header';
+import { Layout } from './layout';
+import { ListTags } from './ListTags';
+import { Pen } from './CodePen';
+import { Posts } from './posts';
+import { Seo } from './seo';
+import { SiteMap } from './sitemap';
+import { Tags } from './Tags';
+import { PassWords } from './PassWords';
+import { ImageHeader } from './imageHeader';
 
 export {
   ExternalLink,
@@ -18,4 +20,6 @@ export {
   Seo,
   SiteMap,
   Tags,
+  PassWords,
+  ImageHeader,
 };
