+++ src/components/layout.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useStaticQuery, graphql, Link } from 'gatsby';
 import styled from 'styled-components';
-import Header from './imageHeader';
+import { ImageHeader } from './';
 import './layout.css';
 
 const StyledMainContainer = styled.div`
@@ -28,7 +28,7 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
 
   return (
     <>
-      <Header />
+      <ImageHeader />
       <StyledMainContainer>
         <main>{children}</main>
         <StyledFooter>
@@ -43,4 +43,4 @@ const Layout = ({ children }: { children: React.ReactNode }) => {
   );
 };
 
-export default Layout;
+export { Layout };
