--- src/components/pass-words/index.ts
@@ -1,1 +1,0 @@
-export * from './pass-words';
