+++ tsconfig.json
@@ -10,7 +10,7 @@
       "esnext"
     ] /* Specify a set of bundled library declaration files that describe the target runtime environment. */,
     "jsx": "react" /* Specify what JSX code is generated. */,
-    "jsxImportSource": "react" /* Specify module specifier used to import the JSX factory functions when using `jsx: react-jsx*`.*/,
+    // "jsxImportSource": "react" /* Specify module specifier used to import the JSX factory functions when using `jsx: react-jsx*`.*/,
 
     /* Modules */
     "module": "esnext" /* Specify what module code is generated. */,
